server/requirements.txt
python-telegram-bot==20.7
cryptography==41.0.7
flask
flask-cors